Work on #65 [skip ci]
